% Generated by roxygen2 (4.0.1): do not edit by hand
\name{affinity}
\alias{affinity}
\title{compute transcription factor binding site affinity}
\usage{
affinity(pwm, seq, Rmax = NULL, lambda = 0.7, pseudo.count = 1,
  gc.content = 0.5, slide = FALSE, both.strands = TRUE)
}
\arguments{
\item{pwm}{position specific count matrix with 4 rows: A, C, G, T}

\item{seq}{the sequence}

\item{Rmax}{the maximal affinity times the activity of the factor by
default set to \code{exp(0.584 * ncol(pwm) - 5.66)}}

\item{lambda}{scaling parameter for the Berg and von Hippel mismatch energy}

\item{pseudo.count}{pseudo count to add to the PWM}

\item{gc.content}{GC content for the background model}

\item{slide}{if true the affinities for each sequence position will
be returned}

\item{both.strands}{scan on both strands (default is TRUE for DNA) or only on the forward
strand (for instance for RNA)}
}
\value{
if slide is false, the function returns to overall affinity, otherwise
 for each of the \code{nchar(seq) - ncol(pwm) + 1} sequence positions
 the score of the site is returned.
 If the sequence contains gap characters "-" then the positions will be
 removed and if \code{slide} is true the positions with gaps will be
 filled with NA. This is usefull if you want to compare polymorphisms
 in promoter sequences.
}
\description{
The TRAP method to predict transcription factor binding affinity to DNA
sequences (Roider et al 2007). Default is to compute the affinity for
the whole region by summing over all sites. If site specific affinities
are of interest set \code{slide} to true.
}
\examples{
pwm = matrix(c(5, 4, 3, 1, 10, 12, 5, 3, 3, 5, 3, 10), nrow=4)
seq = "ACTGACGTGTGCACACGATGCTAGCTG"
affinity(pwm, seq)
}
\references{
Roider, H. G.; Kanhere, A.; Manke, T. & Vingron, M. Predicting transcription factor affinities to DNA from a biophysical model. Bioinformatics, 2007, 23, 134-141
}

