% Generated by roxygen2 (4.0.1): do not edit by hand
\name{evaluate.ranking}
\alias{evaluate.ranking}
\title{Evaluate Ranking}
\usage{
evaluate.ranking(ranking, seq.name2matrix, roc.plot = FALSE, add = FALSE,
  col = "black", predictors = c("ratio", "min", "prod", "rawratio"))
}
\arguments{
\item{ranking}{}

\item{seq.name2matrix}{}

\item{roc.plot}{}

\item{add}{}

\item{col}{}

\item{predictors}{}
}
\value{
Raw AUC value as well as ratio, min and product
}
\description{
Evaluate Ranking
}
\details{
Unknown
}

