% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{exactgevparams}
\alias{exactgevparams}
\title{Precalculated generalized extreme value distribution parameters for
the JASPAR and TRANSFAC database matrices. Calculated for a specific set of
sequence lengths (100, 200, 500, 600, 800, 1000, 2000 and 5000 bp)}
\format{A list where each element is a matrix of parameters (columns) for
each pwm (rows). The list index is the length of sequence for which the
parameters were fit.}
\description{
Precalculated generalized extreme value distribution parameters for
the JASPAR and TRANSFAC database matrices. Calculated for a specific set of
sequence lengths (100, 200, 500, 600, 800, 1000, 2000 and 5000 bp)
}
\keyword{datasets}

