% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.affinity.cutoff}
\alias{get.affinity.cutoff}
\title{determine significance threshold for tRap affinities}
\usage{
get.affinity.cutoff(matrix.name, pwm, cutoff)
}
\arguments{
\item{pwm}{the position weight matrix}

\item{matrix.name}{name of the transfac matrix}

\item{pvalue.cutoff}{significance threshold for the TFBS hits}
}
\value{
affinity cutoff that corresponds to the given P-value
}
\description{
Get affinity cutoff values that corresponds to a given P-value
}

