% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{gevparams}
\alias{gevparams}
\title{parameters for the fitted extreme value distributions for
affinities of each transfac matrix}
\usage{
data(gevparams)
}
\description{
parameters for the fitted extreme value distributions for
affinities of each transfac matrix
}
\details{
parameters for the fitted extreme value distributions for
affinities of each transfac matrix fitted from human promoter sequences
}
\references{
Manke et al.
}

