% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{jaspar}
\alias{jaspar}
\title{The JASPAR transcription factor database in the format expected by tRap}
\format{A list where each element is a single JASPAR matrix and is named by
the JASPAR accession number}
\source{
http://jaspar.genereg.net/html/DOWNLOAD/JASPAR_CORE/pfm/nonredundant/pfm_all.txt
}
\description{
The JASPAR transcription factor database in the format expected by tRap
}
\references{
Sandelin A, Alkema W, Engstrom P, Wasserman WW, Lenhard B.
JASPAR: an open-access database for eukaryotic transcription factor binding profiles.
Nucleic Acids Res. 2004 Jan 1;32(Database issue):D91-4.
}
\keyword{datasets}

