% Generated by roxygen2 (4.0.1): do not edit by hand
\name{local.paffinity}
\alias{local.paffinity}
\title{Calculate local p-values}
\usage{
local.paffinity(affnt, pwm, seq, Rmax = NULL, lambda = 0.7,
  pseudo.count = 1, gc.content = 0.5, window.size = NULL,
  window.offset = 1)
}
\arguments{
\item{affnt}{}

\item{pwm}{}

\item{seq}{}

\item{Rmax}{}

\item{lambda}{}

\item{pseudo.count}{}

\item{gc.content}{}

\item{window.size}{}

\item{window.offset}{}
}
\value{
p-value
}
\description{
Compute local affinity p-values
}
\details{
Use an empirical distribution to calculate a p-value for a given affinity
value using a set of sequences that are passed in.
}

