% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize.pwm}
\alias{normalize.pwm}
\title{normalize a PWM}
\usage{
normalize.pwm(matrix)
}
\arguments{
\item{matrix}{position specific count matrix}
}
\value{
normalized count matrix
}
\description{
Normalize a PWM such that the columns sum to one to use with the
\pkg{seqLogo} package.
}

