% Generated by roxygen2 (4.0.1): do not edit by hand
\name{paffinity}
\alias{paffinity}
\title{compute p-values for affinity values}
\usage{
paffinity(affinity, matrix.name, len, params = NULL, no.exact = FALSE)
}
\arguments{
\item{affinity}{trap affinity value}

\item{matrix.name}{name of the transfac matrix (ignored if params not NULL)}

\item{len}{length of the region used to compute the affinity}

\item{params}{if params is NULL (default), precomputed parameters
will be loaded otherwise params are c(shape, scale, loc) given to
the generalized extreme value distribution}

\item{no.exact}{do not use the parameters that were learned for sequences of the exact same length as \code{len}}
}
\value{
p-value of the affinity score(s)
}
\description{
Use generalized extreme value distributions fitted for each matrix to
compute p-values of trap affinities.
}
\details{
given the affinity compute the p value for the matrix and the length len
if params is NULL, precomputed parameters will be loaded other wise params
are c(shape, scale, loc) given to the generalized extreme value distribution
}
\examples{
\dontrun{
matrices = read.transfac.dir(dir)
seq = read some fasta or get from BSgenome package
matrix.name = "AP1.."
af = affinity(matrices[[matrix.name]], seq)
paffinity(af, matrix.name, 200)}
}
\references{
Thomas Manke et.al.
}

