% Generated by roxygen2 (4.0.1): do not edit by hand
\name{R_affinity_multi}
\alias{R_affinity_multi}
\title{Calculate the TRAP affinity for each position along one or more sequences}
\usage{
R_affinity_multi(pwm, motiflength, sequence, seqlength, Rmax, lambda,
  bothstrands)
}
\arguments{
\item{pwm}{The position weight matrix}

\item{motiflength}{The length of the PWM}

\item{sequence}{a list of one or more genomic sequences}

\item{seqlength}{the lengths of the sequences}

\item{Rmax}{rmax}

\item{lambda}{lambda}

\item{bothstrands}{should the affinity be calcuated on both strands (DNA) or just one (RNA)}
}
\value{
a list where each element is a numeric vector containing the
scores along the length of a single sequence.
}
\description{
Calculate the TRAP affinity for each position along one or more sequences
}

