% Generated by roxygen2 (4.0.1): do not edit by hand
\name{R_affinity_sum_multi}
\alias{R_affinity_sum_multi}
\title{Calculate the TRAP affinity for one or more sequences (one score per sequence)}
\usage{
R_affinity_sum_multi(pwm, motiflength, sequence, seqlength, Rmax, lambda,
  bothstrands)
}
\arguments{
\item{pwm}{The position weight matrix}

\item{motiflength}{The length of the PWM}

\item{sequence}{a list of one or more genomic sequences}

\item{seqlength}{the lengths of the sequences}

\item{Rmax}{rmax}

\item{lambda}{lambda}

\item{bothstrands}{should the affinity be calcuated on both strands (DNA) or just one (RNA)}
}
\value{
a numeric vector of the scores for each sequence
}
\description{
Calculate the TRAP affinity for one or more sequences (one score per sequence)
}

