% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.jaspar}
\alias{read.jaspar}
\title{read a jaspar matrix}
\usage{
read.jaspar(file)
}
\arguments{
\item{file}{jaspar file for all matrices}
}
\value{
returns a list of matrices with four rows for A, C, G, T
}
\description{
Load a PWM matrix from Jaspar
}
\details{
Jaspar file for all matrices should have the format as provided by
jaspar: http://jaspar.genereg.net/html/DOWNLOAD/jaspar_CORE/non_redundant/all_species/matrix_only/matrix_only.txt
}
\author{
Matthias Heinig
}
\seealso{
\code{\link{read.transfac.dir}}
}

