% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.pscm}
\alias{read.pscm}
\title{read a PSCM}
\usage{
read.pscm(file)
}
\arguments{
\item{file}{a file containing position specific count matrices}
}
\value{
returns a list of matrices with four rows for A, C, G, T
}
\description{
Load a generic position-specific count matrix
}
\author{
Matthias Heinig
}
\seealso{
\code{\link{read.transfac.dir}}
}

