% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.transfac.dir}
\alias{read.transfac.dir}
\title{read a directory of transfac matrices}
\usage{
read.transfac.dir(dir)
}
\arguments{
\item{dir}{directory of transfac matrices}
}
\value{
read.transfac returns a list of matrices with four rows for A, C, G, T
}
\description{
Load a whole directory of TRANSFAC matrices
}
\details{
read.transfac.dir returns a list of matrices read from a directory,
named with the filenames of the matrix files.
}
\author{
Matthias Heinig
}
\seealso{
\code{\link{read.transfac}} to read a single matrix
}

