% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.transfac}
\alias{read.jaspar}
\alias{read.transfac}
\alias{read.transfac.dir}
\title{read a transfac matrix}
\usage{
read.transfac(file)
}
\arguments{
\item{file}{transfac matrix / jaspar file for all matrices}

\item{dir}{directory of transfac matrices}
}
\value{
read.transfac returns a list of matrices with four rows for A, C, G, T
read.transfac.dir returns a list of matrices read from a directory,
named with the filenames of the matrix files.
}
\description{
Load a PWM matrix from transfac
}
\details{
Jaspar file for all matrices should have the format as provided by
jaspar: http://jaspar.genereg.net/html/DOWNLOAD/jaspar_CORE/non_redundant/all_species/matrix_only/matrix_only.txt
}
\author{
Matthias Heinig
}
\seealso{
\code{\link{affinity}}
}

