% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scan.sequence}
\alias{scan.sequence}
\title{identify TFBS hits in the sequence(s)}
\usage{
scan.sequence(seq, pwm, matrix.name, pvalue.cutoff, affinity.cutoff = NULL)
}
\arguments{
\item{seq}{character vector with sequences to scan}

\item{pwm}{the position weight matrix}

\item{matrix.name}{name of the transfac matrix}

\item{pvalue.cutoff}{significance threshold for the TFBS hits}

\item{affinity.cutoff}{score threshold (default NULL)}
}
\value{
a matrix with 4 columns: seq, start, end, affinity containing the positions of significant hits
}
\description{
Identify TFBS exceeding the given P-value (or affinity) cutoff
}

